requirements.txt
# package to use FEDGCAN.
absl-py==2.2.2
astunparse==1.6.3
audioread==3.0.1
ca-certificates==2025.2.25
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
common==0.1.2
cycler==0.12.1
decorator==5.2.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
joblib==1.4.2
keras==2.10.0
keras-preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
libffi==3.4.4
librosa==0.9.1
llvmlite==0.41.1
markdown==3.7
markupsafe==2.1.5
matplotlib==3.5.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
openssl==3.0.16
opt-einsum==3.4.0
packaging==25.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pooch==1.8.2
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyparsing==3.1.4
python==3.8.20
python-dateutil==2.9.0.post0
pyyaml==6.0
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rsa==4.9.1
scikit-learn==1.0.2
scipy==1.10.1
setuptools==75.1.0
six==1.17.0
soundfile==0.13.1
sqlite==3.45.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.63.0
typing-extensions==4.13.2
urllib3==2.2.3
vc==14.42
vs2015_runtime==14.42.34433
werkzeug==3.0.6
wheel==0.44.0
wrapt==1.17.2
zipp==3.20.2
